Sekiban.Web.OpenApi: Move extension methods to 'extentions' subfolder.
