+++ internalUsages/MultiTenant.WebApi/Program.cs
@@ -2,7 +2,7 @@ using MultiTenant.WebApi;
 using Sekiban.Core.Dependency;
 using Sekiban.Infrastructure.Cosmos;
 using Sekiban.Web.Dependency;
-using Sekiban.Web.OpenApi;
+using Sekiban.Web.OpenApi.Extensions;
 var builder = WebApplication.CreateBuilder(args);
